requirements.txt
fastapi
uvicorn
groq
sqlalchemy
pgvector
psycopg[binary]
pypdf
streamlit 
phidata
google-generativeai
duckduckgo-search